move variable and step output to common outputs
